+++ src/Command/UpdateTranslationsConsoleCommand.php
@@ -16,6 +16,10 @@ use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 
+if (!defined('_PS_VERSION_')) {
+    exit;
+}
+
 class UpdateTranslationsConsoleCommand extends Command
 {
     public function __construct()