+++ src/Handler/Certificate/ApplePayDirectCertificateHandler.php
@@ -17,6 +17,10 @@ use Mollie\Factory\ModuleFactory;
 use Mollie\Handler\Certificate\Exception\ApplePayDirectCertificateCreation;
 use Mollie\Utility\FileUtility;
 
+if (!defined('_PS_VERSION_')) {
+    exit;
+}
+
 class ApplePayDirectCertificateHandler implements CertificateHandlerInterface
 {
     const FILE_NAME = 'ApplePayDirectCertificateHandler';