+++ src/Service/ConfigFieldService.php
@@ -19,6 +19,10 @@ use Mollie\Factory\ModuleFactory;
 use Mollie\Repository\CountryRepository;
 use Mollie\Utility\EnvironmentUtility;
 
+if (!defined('_PS_VERSION_')) {
+    exit;
+}
+
 class ConfigFieldService
 {
     /**