+++ src/Service/PaymentMethodService.php
@@ -50,6 +50,10 @@ use PrestaShopDatabaseException;
 use PrestaShopException;
 use Tools;
 
+if (!defined('_PS_VERSION_')) {
+    exit;
+}
+
 class PaymentMethodService
 {
     /**